Removed Albite.Core.UnitTest.
It turns out that Assert.ThrowsException<> does behave correctly.

However, there appeared to be a problem for that the exception is
called by the debugger before it is called by Assert.ThrowsException<>.
I observed the same behaviour when the method is in a different
*compiled* assembly (e.g. Albite.Core from Albite.Serialization).

So, it is not needed at all.